requirements.txt
dj-database-url==0.5.0
Django==2.0
django-annoying==0.10.4
django-bootstrap4==0.0.6
django-heroku==0.3.1
django-registration==2.4.1
djangorestframework==3.8.2
gunicorn==19.8.1
pytz==2018.4
whitenoise==3.3.1